requirements.txt
opencv-python==4.8.1.78
tensorflow==2.13.0
mtcnn==0.1.1
numpy==1.24.3
pandas==1.5.3
Pillow==10.0.1
scikit-learn==1.3.0
sqlite3
tkinter
openpyxl==3.1.2
matplotlib==3.7.2
seaborn==0.12.2